pyproject.toml
[tool.poetry]
name = "pil1-2324-2"
version = "0.1.0"
description = ""
authors = ["bopo <ibopo@126.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django-widget-tweaks = "^1.5.0"
django = "^5.0.7"
channels = "^4.1.0"
django-crispy-forms = "^2.3"
crispy-bootstrap5 = "^2024.2"
django-filter = "^24.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
